all tests in and working
