+++ a11y-checker/package.json
@@ -18,6 +18,7 @@
     "babel-eslint": "^10.0.1",
     "babel-jest": "^24.5.0",
     "jest": "^24.5.0",
+    "jest-puppeteer": "^4.1.1",
     "puppeteer-request-spy": "^1.3.0"
   }
 }
