+++ a11y-checker/src/__mocks__/pass.html
@@ -1,5 +1,6 @@
-<html>
+<html lang="en">
+  <head><title>Test!</title></head>
   <body>
-    <main class="App">Test</main>
+    <main class="App"><h1>Test!</h1>Test</main>
   </body>
 </html>
