+++ a11y-checker/src/__tests__/lib/requestScanFail.test.js
@@ -8,29 +8,7 @@ const matcher = (testee, keyword) => {
   return testee.indexOf(keyword) > -1;
 };
 
-it("Return true when html passes axe tests", async () => {
-  const domain = "https://www.example.org";
-  const htmlPath = path.resolve(__dirname, "../../__mocks__/pass.html");
-  const fileResult = await getFile(htmlPath);
-  const htmlResponseFaker = new ResponseFaker(domain, {
-    status: 200,
-    contentType: "text/html",
-    body: fileResult.toString("utf8")
-  });
-
-  const requestInterceptor = new RequestInterceptor(matcher);
-  requestInterceptor.addFaker(htmlResponseFaker);
-  page.setRequestInterception(true);
-  page.on("request", requestInterceptor.intercept.bind(requestInterceptor));
-
-  const result = await requestScan(domain, {
-    browser,
-    page
-  });
-
-  expect(result).toEqual(true);
-});
-it("Return false when html fails axe tests", async () => {
+it("Returns false when html fails axe tests", async () => {
   const domain = "https://www.example.org";
   const htmlPath = path.resolve(__dirname, "../../__mocks__/fail.html");
   const fileResult = await getFile(htmlPath);
@@ -50,5 +28,5 @@ it("Return false when html fails axe tests", async () => {
     page
   });
 
-  expect(result).notToEqual(true);
+  expect(result).not.toEqual(0);
 });
